Refactor new code into a KSearchField widget
